use the isset test
That test on line 2: "if(!isset($_POST['submit']))" checks to see if the form was submitted. You should only set all those variables and send the message if the test is true. The way you had it, if I didn't come to this page via submit, it still tries to send an email.